Minor fixes: derive_params updates, cleanup unused imports
- derive_params.py: expanded parameter derivation for sweep automation
- tune_hbm_buffer.py: improved tuning script with port conflict handling
- Remove unused imports in iris.py, tracing/events.py, x/core.py

+++ iris/ops/all_gather_matmul_hbm_buffer.py
@@ -219,7 +219,7 @@ def _hbm_buffer_all_gather_matmul_kernel(
             acc = acc + bias_val[:, None]
 
         c = acc.to(C.type.element_ty)
-        C_ptrs = C + rm[:, None] * stride_cm + rn[None, :] * stride_cn
+        C_ptrs = C + rm.to(tl.int64)[:, None] * stride_cm + rn[None, :] * stride_cn
         c_mask = (rm[:, None] < M) & (rn[None, :] < N)
         tl.store(C_ptrs, c, mask=c_mask, cache_modifier=".wt")
